### 2023-07-03 #5617 @editorialbot
[PRE REVIEW]: hdlib: A Python library for designing Vector-Symbolic Architectures
**Submitting author:** @cumbof (<a href="http://orcid.org/0000-0003-2920-5838">Fabio Cumbo</a>)
**Repository:** https://github.com/cumbof/hdlib
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.14
