### 2023-07-03 #5617 @danielskatz
👋 @cumbof - thanks for this submission. Can I suggest that you add a bit to the README about what the software is for, similar to some of the paper introduction?

I'm also going to ask JOSS editors to review this for scope, due to the relatively small amount of code and the short commit history. The editors will discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.
