### 2023-07-03 #5617 @cumbof
Thanks @danielskatz for your processing this submission so quickly!
I'm going to update the `README` pretty soon (in the next few hours).

Just want to highlight the point that the idea of building this library started from an attempt of generalising [chopin2](https://github.com/cumbof/chopin2), a software that implements a supervised classification model based on the hyperdimensional computing (HD) paradigm. It also implements a backward variable selection technique as the very first attempt of building a feature selection method with a HD approach.

We generalised everything and we also implemented a series of abstractions that help building vector-symbolic architectures very easily. With this library, we have been able to reimplement `chopin2` from scratch drastically improving its running time (see [https://github.com/cumbof/hdlib/blob/main/examples/chopin2.py](https://github.com/cumbof/hdlib/blob/main/examples/chopin2.py)). We are planning to release a new version of `chopin2` that makes use of `hdlib` pretty soon.

Hope in a positive response from the JOSS Editors
