### 2023-07-07 0492d7b
Fix the retraining process in model.Model.predict
