+++ CITATION.cff
@@ -4,7 +4,7 @@ authors:
   - family-names: Cumbo
     given-names: Fabio
     orcid: https://orcid.org/0000-0003-2920-5838
-title: "cumbof/hdlib: hdlib v0.1.8"
+title: "cumbof/hdlib: hdlib v0.1.9"
 version: 0.1.8
-date-released: 2023-06-30
+date-released: 2023-07-07
 url: https://github.com/cumbof/hdlib
