### 2023-07-10 37ccc6a
Update model.py
Add error_rate function; Compute the error rate before retraining the classification model
