### 2023-07-13 d8316dd
Update model.py
The `model.Model.stepwise_regression` function now report the importance corresponding to the best score; The `model.Model._init_fit_predict` function uses `average="weighted"` for computing a score different from the accuracy to account for label imbalance.
