### 2023-07-13 4c2d37f
Update chopin2.py
It now computes different scores on the resulting predictions, prints the list of selected features based on the best score, and finally reports the confusion matrices.
