+++ examples/chopin2.py
@@ -1,20 +1,17 @@
 #!/usr/bin/env python3
-"""
-Implementation of chopin2 ML model with hdlib
-https://github.com/cumbof/chopin2
-"""
+"""Implementation of chopin2 ML model with hdlib."""
 
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
 __version__ = "1.1.0"
-__date__ = "Jun 14, 2023"
+__date__ = "Jul 13, 2023"
 
 import argparse as ap
 import os
 import statistics
 import time
 
-from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
+from sklearn.metrics import accuracy_score, confusion_matrix, f1_score, precision_score, recall_score
 from tabulate import tabulate
 
 from hdlib.parser import load_dataset, percentage_split
@@ -22,10 +19,12 @@ from hdlib.model import Model
 
 
 def read_params():
-    """
-    Read and test input arguments
+    """Read and test the input arguments.
 
-    :return:        The ArgumentParser object
+    Returns
+    -------
+    argparse.ArgumentParser
+        The ArgumentParser object
     """
 
     p = ap.ArgumentParser(
@@ -113,8 +112,12 @@ def read_params():
 
 
 def chopin2():
-    """
-    The ML model implemented in chopin2 is described in the following scientific paper:
+    """Domain-agnostic supervised learning with hyperdimensional computing.
+    
+    Notes
+    -----
+    chopin2 is available in GitHub at https://github.com/cumbof/chopin2 and its ML model is described
+    in the following scientific article:
 
     Cumbo F, Cappelli E, Weitschek E.
     A brain-inspired hyperdimensional computing approach for classifying massive dna methylation data of cancer
@@ -145,75 +148,20 @@ def chopin2():
     
     _, features, content, classes = load_dataset(args.input, sep=args.fieldsep)
 
-    print("Points: {}; Features {}; Classes {}".format(len(content), len(features), len(set(classes))))
+    class_labels = sorted(list(set(classes)))
+
+    print("Points: {}; Features {}; Classes {}".format(len(content), len(features), len(class_labels)))
 
     # Initialise the model
     print("Building the HD model")
     
     model = Model(size=args.dimensionality, levels=args.levels, vtype="bipolar")
 
-    if not args.feature_selection:
-        # Fit the model
-        model.fit(content, classes)
-
-        if args.kfolds > 0:
-            # Predict in cross-validation
-            print("Cross-validating model with {} folds".format(args.kfolds))
-
-            predictions = model.cross_val_predict(
-                content,
-                classes,
-                cv=args.kfolds,
-                distance_method="cosine",
-                retrain=args.retrain,
-                n_jobs=args.nproc
-            )
-
-        elif args.test_percentage > 0.0:
-            # Predict with a percentage-split
-            print("Percentage-split: training={}% test={}%".format(100.0 - args.test_percentage, args.test_percentage))
-
-            test_indices = percentage_split(len(content), args.test_percentage)
-
-            predictions = [
-                model.predict(
-                    test_indices,
-                    distance_method="cosine",
-                    retrain=args.retrain,
-                    stop_if_worse=True
-                )
-            ]
-
-        # For each prediction, compute the accuracy, f1, precision, and recall
-        accuracy_scores = list()
-        f1_scores = list()
-        precision_scores = list()
-        recall_scores = list()
-
-        retraining_iterations = list()
-
-        for y_indices, y_pred, retrainings in predictions:
-            y_true = [label for position, label in enumerate(classes) if position in y_indices]
-
-            accuracy_scores.append(accuracy_score(y_true, y_pred))
-            f1_scores.append(f1_score(y_true, y_pred), average="micro")
[…]
