+++ hdlib/__init__.py
@@ -1,4 +1,4 @@
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
-__version__ = "0.1.10"
-__date__ = "Jul 10, 2023"
+__version__ = "0.1.11"
+__date__ = "Jul 13, 2023"
