### 2023-07-13 #5617 @cumbof
@danielskatz wondering if there is any news about this.

Also note that `hdlib` has been bumped to version 0.1.11 in the meantime.
At the time of the submission last week, its version was 0.1.8. Wondering if this info must also be updated somehow.
