### 2023-07-16 6af6152
Update model.py
Add checks on input parameters and raise exceptions
