### 2023-07-16 e7bb626
Update chopin2.py
Report Accuracy, F1, Precision, Recall, MCC, and the confusion matrix for each of the folds
