+++ CITATION.cff
@@ -4,7 +4,7 @@ authors:
   - family-names: Cumbo
     given-names: Fabio
     orcid: https://orcid.org/0000-0003-2920-5838
-title: "cumbof/hdlib: hdlib v0.1.11"
-version: 0.1.11
-date-released: 2023-07-13
+title: "cumbof/hdlib: hdlib v0.1.12"
+version: 0.1.12
+date-released: 2023-07-16
 url: https://github.com/cumbof/hdlib
