### 2023-07-16 #5617 @cumbof
@danielskatz `hdlib` version is now 0.1.12 and, based on my `cloc`, the number of Python lines (code) in the `main` branch is 1,272.

Now, since the JOSS documentation reports that:

> Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, thos under 300 LOC will be desk rejected.

Do you still think that asking the Editors to review this submission for the scope is still necessary?
