### 2023-07-17 8ce8ff4
Fix the retraining process in predict()
