### 2023-07-21 #5617 @cumbof
That's a great news, thanks @danielskatz and JOSS Editors for considering our submission!
I'll suggest a few potential reviewers asap.
