### 2023-07-22 #5617 @cumbof
@danielskatz here is a list of a few potential reviewers:
- Mohsen Imani ([moimani](https://github.com/moimani));
- Abbas Rahimi ([abbas-rahimi](https://github.com/abbas-rahimi));
- Ross Gayler ([rgayler](https://github.com/rgayler));
- Anıl Tuncel ([anilbey](https://github.com/anilbey));
- Mahfuzur Rahman ([mahfuz05062](https://github.com/mahfuz05062));
- Ritika Giri ([ritika-giri](https://github.com/ritika-giri)).

I would strongly recommend the first three as experts in the field of hyperdimensional computing (a.k.a. vector-symbolic architectures) and the other three as per their strong expertise with Python.

Please note that the first three are not in [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review.
