### 2023-07-24 #5617 @rgayler
@danielskatz thanks for the invitation to review, but I must refuse. Possibly uniquely among those who know about hyperdimensional computing I have never used Python, so am probably uniquely unqualified to review hdlib as software.

While reading the reviewing guidelines in order to decide whether I could adequately review the paper I noticed that the guidelines require related software to be cited. The authors (@cumbof ) should look at the following software to determine the degree of overlap:

- [](https://github.com/hyperdimensional-computing/torchhd)
- [](https://github.com/Adam-Vandervorst/PyBHV)
- [](https://github.com/UCSD-SEELab/openhd)
- [](https://pypi.org/project/hdtorch/)
