### 2023-07-24 #5617 @danielskatz
@cumbof

> Thanks for pointing that out @rgayler. I'm not sure I'm allowed to change anything in the manuscript at this point. I'll be waiting for comments from @danielskatz and reviewers

You can make changes before we start the review, so if you can do this in a day or two, please do. And in fact, the review is fairly iterative and this can happen at a detailed level, so it's usually fine to make changes during the review process as well, as long as reviewers can see what you have done (you say so as a comment in this or the review thread, once we get there)
