### 2023-07-25 5d2bd05
Updating manuscript
Add references to existing libraries
