+++ paper/paper.bib
@@ -198,3 +198,35 @@
   publisher={Elsevier},
   doi={10.1016/j.future.2020.04.005}
 }
+
+@misc{heddes2023torchhd,
+  title={Torchhd: An Open Source Python Library to Support Research on Hyperdimensional Computing and Vector Symbolic Architectures}, 
+  author={Mike Heddes and Igor Nunes and Pere Vergés and Denis Kleyko and Danny Abraham and Tony Givargis and Alexandru Nicolau and Alexander Veidenbaum},
+  year={2023},
+  eprint={2205.09208},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.2205.09208}
+}
+
+@article{kang2022openhd,
+  title={OpenHD: A gpu-powered framework for hyperdimensional computing},
+  author={Kang, Jaeyoung and Khaleghi, Behnam and Rosing, Tajana and Kim, Yeseong},
+  journal={IEEE Transactions on Computers},
+  volume={71},
+  number={11},
+  pages={2753--2765},
+  year={2022},
+  publisher={IEEE},
+  doi={10.1109/TC.2022.3179226}
+}
+
+@misc{simon2022hdtorch,
+  title={HDTorch: Accelerating Hyperdimensional Computing with GP-GPUs for Design Space Exploration}, 
+  author={William Andrew Simon and Una Pale and Tomas Teijeiro and David Atienza},
+  year={2022},
+  eprint={2206.04746},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.2206.04746}
+}
