### 2023-07-25 #5617 @cumbof
@danielskatz I've updated the manuscript by adding three references and a couple of short sentences.
Also note that a few days ago I bumped the version of the library to 0.1.13
