### 2023-07-27 42bdbc2
Update Vector.dist
Convert cosine similarity into a distance measure
