### 2023-07-27 c5f7989
Fix fit, error_rate, and predict
(i) `model.Model.fit` now correctly generates both bipolar and binary level vectors; (ii) `model.Model.predict` and `model.Model.error_rate` are now compatible with all the supported distance metrics (euclidean, hamming, and cosine).
