+++ hdlib/__init__.py
@@ -1,4 +1,4 @@
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
-__version__ = "0.1.13"
-__date__ = "Jul 17, 2023"
+__version__ = "0.1.14"
+__date__ = "Jul 26, 2023"
