+++ CITATION.cff
@@ -4,7 +4,7 @@ authors:
   - family-names: Cumbo
     given-names: Fabio
     orcid: https://orcid.org/0000-0003-2920-5838
-title: "cumbof/hdlib: hdlib v0.1.13"
-version: 0.1.13
-date-released: 2023-07-17
+title: "cumbof/hdlib: hdlib v0.1.14"
+version: 0.1.14
+date-released: 2023-07-26
 url: https://github.com/cumbof/hdlib
