### 2023-07-27 32d0142
Update space.py
Fix default threshold in Space.find and Space.find_all
