### 2023-07-27 #5617 @mahfuz05062
I think I already showed my interest in reviewing this earlier.

On Thu, Jul 27, 2023 at 9:11 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @ritika-giri <https://github.com/ritika-giri>, @Anibey
> <https://github.com/Anibey>, @mahfuz05062 <https://github.com/mahfuz05062>
> - would one of you be willing to review this submission for JOSS?
>
