### 2023-07-27 #5617 @cumbof
> 👋 @ritika-giri, @Anibey, @mahfuz05062 - would one of you be willing to review this submission for JOSS?

@danielskatz wrong mention? You were supposed to mention @anilbey I guess (missing "l")
