### 2023-08-15 cumbof/hdlib#4 @anilbey
API method docs
Hello,

Could you generate the API method documentation as well? It should be fairly straightforward for this repository since each method already has a detailed docstring in hdlib.

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
