### 2023-08-15 cumbof/hdlib#5 @anilbey
Automated tests are missing
Hi,

Could you automate the tests available at the test module?

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
