### 2023-08-15 cumbof/hdlib#3 @cumbof
Thanks for reporting this issue @anilbey

I've just updated the documentation by importing `bind` and `bundle` from `hdlib.arithmetic`.
Also, the `IndexError` comes out because there are no vectors called `DOL` in the space (it must refer to `USD` actually). Btw, everything's fixed now.
I decided to maintain the vector's name as `DOL` to make it synced with the unit test in `test/test.py`.

Feel free to close this issue if the problem is solved.
Thanks!
