### 2023-08-15 cumbof/hdlib#4 @cumbof
For sure it would be very easy. I'll work on that in the next few hours.
Thanks for the advice @anilbey
