### 2023-08-15 cumbof/hdlib#5 @anilbey
Hello @cumbof, implementing GitHub Actions or another third-party CI would be ideal. If that's not feasible, providing documentation for manual testing steps would also be okay. The goal is for contributors to easily run tests locally and ensure their updates don't disrupt existing functionality.
