### 2023-08-15 cumbof/hdlib#6 @anilbey
sample data for running the examples/chopin2.py script
Hi again @cumbof

I couldn't run the examples/chopin2.py script.

Could you provide a sample input data that is in the expected format?
Additionally, if you could provide a method or a script to run the algorithm on this dataset with recommended parameters, that would be very helpful.

Thanks in advance!

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
