### 2023-08-20 cumbof/hdlib#7 @mahfuz05062
Language / Typo Suggestions
Please go over the text again carefully and correct any typos / grammatical mistakes. A couple of examples:

- Line 36 => Replace 'It does not only consolidate ...' with 'It not only consolidates ...'
- Line 60 => Replace 'then' with 'than'

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
