### 2023-08-20 cumbof/hdlib#8 @mahfuz05062
Elaborate Supports and Contribution
Can we elaborate on the guidelines for Supports and Contribution, for example, how would a potential contributor get started with the contribution process, and how would a contribution lifecycle look like? This is a good resource: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
