### 2023-08-21 cumbof/hdlib#9 @anilbey
missing import statements in model.py
Hello @cumbof ,

Running static code analysis detected missing import statements in the following functions: `Model::auto_tune`, `Model::stepwise_regression` and  `Model::cross_val_predict`.

```shell
> ruff . --ignore=E501

hdlib/model.py:672:18: F821 Undefined name `os`
hdlib/model.py:785:18: F821 Undefined name `os`
hdlib/model.py:994:18: F821 Undefined name `os`
```

Could you make sure the tests in the test module are covering these 3 functions? Thanks

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5704 review.
