+++ hdlib/__init__.py
@@ -1,3 +1,15 @@
+"""__hdlib__ is a Python 3 library for building Vector-Symbolic Architectures (VSA, a.k.a. Hyperdimensional Computing).
+
+VSA is an emergent computing paradigm that works by combining vectors in a high-dimensional space for representing and 
+processing information. This approach recently shown promise in various domains for dealing with different kind of 
+computational problems, including artificial intelligence, cognitive science, robotics, natural language processing, 
+bioinformatics, medical informatics, cheminformatics, and internet of things among other scientific disciplines.
+
+The library is distributed under the MIT license as a Python package through PyPI and Conda on the _conda-forge_ channel.
+
+Please refer to the official [Wiki](https://github.com/cumbof/hdlib/wiki) for any information about the implemented 
+modules and how to use the library."""
+
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
 __version__ = "0.1.14"
