### 2023-08-22 64f0210
Update docstring and fix import
Add import os
