+++ hdlib/__init__.py
@@ -12,5 +12,5 @@ modules and how to use the library."""
 
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
-__version__ = "0.1.14"
+__version__ = "0.1.15"
 __date__ = "Jul 26, 2023"
