### 2023-08-22 cumbof/hdlib#4 @cumbof
I've just deployed the documentation through GitHub Pages. It is accessible at [https://cumbof.github.io/hdlib/](https://cumbof.github.io/hdlib/).

It refers to the latest version v0.1.15 which I still have to release. It will include the `import os` fix in the `hdlib.model` module and a few additional unit tests.

I'm going to release it as soon as the JOSS review process will be completed so that I can collect any other changes in the v0.1.15.

@anilbey Please, feel free to close the issue if everything looks ok for you.
