### 2023-08-22 cumbof/hdlib#5 @cumbof
In order to answer Issue #8, I've defined a [CONTRIBUTING.md](https://github.com/cumbof/hdlib/blob/main/CONTRIBUTING.md) file with a few directions for whoever wants to contribute to the code development.

One of the steps required to propose a contribution is the following one:

> Run the unit tests to make sure that your changes did not mess up with the existing features;
> ...
>> **Note:** In order to run the unit tests, simply run the following command in your terminal from the project root folder: `python -m unittest test.test`.

@anilbey Again, if you think this could be enough, please feel free to close the issue. Thanks!
