### 2023-08-22 cumbof/hdlib#9 @cumbof
Hi @anilbey, thanks for pointing this out.

I fixed the missing import in `hdlib.model` and I will add a few additional unit tests ASAP.
I'll keep you updated here. Thanks
