### 2023-08-22 cumbof/hdlib#8 @cumbof
Hi @mahfuz05062, thanks for the suggestion.

I've added a [CONTRIBUTING.md](https://github.com/cumbof/hdlib/blob/main/CONTRIBUTING.md) file a few minutes ago.
Both the [README.md](https://github.com/cumbof/hdlib/blob/main/README.md) and the [Support and Contributions](https://github.com/cumbof/hdlib/wiki/Support-and-contributions) section of the Wiki also point to the [CONTRIBUTING.md](https://github.com/cumbof/hdlib/blob/main/CONTRIBUTING.md) file now.

Please, feel free to close this issue if you think I appropriately addressed your point.
Thanks!
