### 2023-08-22 01cbcb1
Update paper.md
Fixing typos / grammatical errors as pointed in #7
