### 2023-08-22 cumbof/hdlib#7 @cumbof
Thanks for highlighting these errors @mahfuz05062. I don't see any other typos / grammatical mistakes.
I've just updated the manuscript and I'm going to rebuild the PDF in the official review thread.

Again, feel free to close the issue if it looks ok for you.
Thanks!
