### 2023-08-22 #5704 @cumbof
This is the new version of the manuscript. Actually the same as the previous one with a couple of fixes on grammatical mistakes as spotted by @mahfuz05062 in cumbof/hdlib#7
