### 2023-08-24 c540409
Update parser.py
load_dataset works with numerical datasets only
