### 2023-08-24 adc4f43
Update test.py
Add more unit tests
