+++ test/test.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python3
-"""
-Unit tests for hdlib
-"""
+"""Unit tests for hdlib."""
 
 import errno
 import os
@@ -13,30 +11,25 @@ import numpy as np
 from sklearn import datasets
 from sklearn.metrics import accuracy_score
 
-# Define the test root directory
-TESTS_DIR = os.path.dirname(os.path.realpath(__file__))
-
 # Define the hdlib root directory
-HDLIB_DIR = os.path.join(os.path.dirname(TESTS_DIR), "hdlib")
+ROOT_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+print(ROOT_DIR)
 
 # This is required to import the functions we need to test
-sys.path.append(HDLIB_DIR)
+sys.path.append(ROOT_DIR)
 
-# Finally import the space, vector, and arithmetic operators
-from space import Space, Vector
-from arithmetic import bundle, bind, permute
-from model import Model
+# Finally import the space, vector, arithmetic operators, model, and parser utilities
+from hdlib.space import Space, Vector
+from hdlib.arithmetic import bundle, bind, permute
+from hdlib.model import Model
+from hdlib.parser import kfolds_split, percentage_split
 
 
 class TestHDLib(unittest.TestCase):
-    """
-    Unit tests for hdlib
-    """
+    """Unit tests for hdlib"""
 
     def test_vector(self):
-        """
-        Unit tests for hdlib/space.py:Vector class
-        """
+        """Unit tests for hdlib/space.py:Vector class"""
 
         # Create a random binary numpy.ndarray
         ndarray = np.random.randint(2, size=10000)
@@ -75,9 +68,7 @@ class TestHDLib(unittest.TestCase):
                 self.assertTrue(np.array_equal(bipolar_vector.vector, pickle_vector.vector))
 
     def test_space(self):
-        """
-        Unit tests for hdlib/space.py:Space class
-        """
+        """Unit tests for hdlib/space.py:Space class"""
 
         # Create a Space object
         space = Space(vtype="bipolar")
@@ -136,13 +127,34 @@ class TestHDLib(unittest.TestCase):
             with self.subTest():
                 self.assertEqual(len(space), len(pickle_space))
 
-    def test_model(self):
-        """
-        Unit tests for hdlib/model.py:Model class
-        """
+    def test_arithmetic(self):
+        """Unit tests for hdlib/arithmetic.py"""
 
-        # Create a model with vector dimensionality 10000, vector type bipolar, and 10 levels
-        model = Model(size=10000, levels=10, vtype="bipolar")
+        # Create two vectors to test the arithmetic functions
+        vector1, vector2 = Vector(), Vector()
+
+        # Element-wise multiplication of vector1 and vector2
+        bind_vector = bind(vector1, vector2)
+
+        with self.subTest():
+            self.assertFalse(all(bind_vector.vector == vector1.vector) and all(bind_vector.vector == vector2.vector))
+
+        # Element-wise sum of vector1 and bind_vector
+        bundle_vector = bundle(vector1, bind_vector)
+
+        with self.subTest():
+            self.assertFalse(all(bundle_vector.vector == vector1.vector) and all(bundle_vector.vector == bind_vector.vector))
+
+        # Rotate bundle_vector by 1 position
+        permute_vector = permute(bundle_vector, rotate_by=1)
+
+        with self.subTest():
+            # The permute function is invertible
+            # Rotating permute_vector by -1 positions will produce bundle_vector again
+            self.assertTrue(all(bundle_vector.vector == permute(permute_vector, rotate_by=-1).vector))
+
+    def test_model(self):
[…]
