+++ hdlib/__init__.py
@@ -13,4 +13,4 @@ modules and how to use the library."""
 __author__ = ("Fabio Cumbo (fabio.cumbo@gmail.com)")
 
 __version__ = "0.1.15"
-__date__ = "Jul 26, 2023"
+__date__ = "Aug 24, 2023"
