### 2023-08-24 cumbof/hdlib#9 @cumbof
@anilbey I've just added a few more unit tests that cover also `Model::stepwise_regression` and `Model::auto_tune`.
`Model::cross_val_predict` was already covered by a unit test.

Feel free to close the issue if you think everything is ok now.
Thanks!
