### 2023-08-25 cumbof/hdlib#6 @cumbof
Hi @anilbey,

I've added a [examples/README.md](https://github.com/cumbof/hdlib/blob/main/examples/README.md) file a few minutes ago. It doesn't add a lot of information but this point is crucial:

> As the original software, the chopin2.py script only accepts numerical datasets in input. The first row must contain the header with the name of the features, while the first column contains the name of the samples and the last column the class labels.
>> **Note:** Given an input matrix M, values in position M(ij) must be numbers, except for the first row, first column, and last column.

I've also updated the `load_dataset()` function in `hdlib.parser` to throw a `ValueError` in case the input dataset contains categorical values.

Could this be the reason why you were not able to run the script?

The following python code extract the `iris` dataset from `scikit-learn` and dump it to a tsv file ready to be used as input to the `chopin2.py` script:

```python
from sklearn import datasets

iris = datasets.load_iris()

with open("~/iris.tsv", "w+") as dataset:
    dataset.write("sample\t{}\tclass\n".format("\t".join(iris.feature_names)))
    labels = iris.target.tolist()

    for pos, point in enumerate(iris.data.tolist()):
        dataset.write("{}\t{}\t{}\n".format(pos, "\t".join([str(value) for value in point]), labels[pos]))
```

Note that the `iris` datasets contains samples distributed over 3 classes. We didn't investigate the performance of the script over a multi-class dataset, so I would suggest to split it into 3 distinct datasets by considering classes 0 and 1, 1 and 2, and 0 and 2 separately in a binary settings.

However, as an example, you can definitely use the dataset as produced by the python snippet above.

You can run `chopin2.py` with the following configuration:

```
python chopin2.py --input ~/iris.tsv --levels 100 --feature-selection backward --kfolds 5 --retrain 10 --nproc 2
```

The number of level vectors `--levels 100` is usually determined on the numerical values in the datasets. Here we have one digit numbers (i.e., in the [0-9] range), so we need 10 vectors to uniquely represent all of them. But you also have to take into account the precision of the number because these are floats, not integers, all with a one digit precision (8.1 is different from 8.2), and if you want to uniquely represent all these shades, you need other 10 vectors for each number, i.e., 10*10 level vectors.

Please note that this estimation could not be the best choice. What if you have the following two numbers, 1.9E-9 and 2E-9. Do you really need to represent them as two distinct vectors? Could you instead represent them as the same vector? There is not an answer here, and it really depends on the dataset. That's the main reason why we implemented the `Model::auto_tune` function. It performs a Parameter Sweep Analysis on the vector dimensionality and the number of level vectors and helps establish the best parameters for your dataset.

I also enable the feature selection `--feature-selection` as backward variable elimination by also cross-validating the classification models with 5-folds `--kfolds 5`.

The retraining process `--retrain 10` acts as an error mitigation technique. The vector representation of classes are built by bundling the vector representation of samples belonging to the same class. The more vectors you add, the more is difficult to retrieve information from the resulting vector. Retraining means that, given a vector representation of a specific class, we iterate over the set of samples belonging to that specific class that we previously used to build the class vector (the training set) and we check whether a specific sample results to be closer to the vector of its class rather than a vector of another class by computing the cosine similarity. If it results closer to a different class, we subtract the vector representation of that sample from the closest class, and we add its vector representation to the vector representation of its class, in order to reduce and increment its signal respectively. And we repeat this process iteratively until the error rate doesn't change anymore or we stop if it gets worse, up to a maximum number of iterations (e.g., 10 in this example).

This process is better explained in [https://doi.org/10.3390/a13090233](https://doi.org/10.3390/a13090233).

Hope this will help.
