### 2023-09-05 cumbof/hdlib#6 @anilbey
Many thanks @cumbof. In addition to the update on the README and the explanation you provided, could you also provide a script in the examples directory that calls the `chopin2.py` on the iris dataset according to the procedure you described above?
