### 2023-09-05 cumbof/hdlib#5 @anilbey
Yes that is sufficient according to JOSS criteria. Thanks @cumbof
