### 2023-09-05 cfc70c4
Update space.py
Simplifying bulk_insert function
