### 2023-09-05 3f3ea47
Add chopin2_iris.sh
Add an example about how to run chopin2.py using the iris.csv dataset from the scikit-learn repository on GitHub
