+++ examples/README.md
@@ -32,3 +32,5 @@ Here is a list of available options:
 | `--uncertainty`       | `5.0`   |           | Uncertainty percentage. Used in conjunction with `--feature-selection` |
 | `--nproc`             | `1`     |           | Make it parallel when possible |
 | `--version`           |         |           | Print the tool version and exit |
+
+> **Note:** The [chopin2\_iris.sh](https://github.com/cumbof/hdlib/blob/main/examples/chopin2_iris.sh) automatically downloads the [iris.csv](https://raw.githubusercontent.com/scikit-learn/scikit-learn/main/sklearn/datasets/data/iris.csv) dataset from the scikit-learn repository on GitHub and run _chopin2.py_ by using 10 level vectors and selecting features through the backward variable elimination technique in 5-folds cross-validation. Also note that this must be intended just as an example about how to run _chopin2.py_ and results strongly depend on the input dataset.
