### 2023-09-05 cumbof/hdlib#6 @cumbof
Hi @anilbey,
I've just added the [chopin2_iris.sh](https://github.com/cumbof/hdlib/blob/main/examples/chopin2_iris.sh) script (commit [https://github.com/cumbof/hdlib/commit/3f3ea47806a8b2b7d9f5f67df0a71be3b5a92d66](https://github.com/cumbof/hdlib/commit/3f3ea47806a8b2b7d9f5f67df0a71be3b5a92d66)) that automatically downloads the [iris.csv](https://raw.githubusercontent.com/scikit-learn/scikit-learn/main/sklearn/datasets/data/iris.csv) dataset from the scikit-learn repository on GitHub and runs [chopin2.py](https://github.com/cumbof/hdlib/blob/main/examples/chopin2.py).

Please let me know if it looks ok for you.
Thanks again
