### 2023-09-08 cumbof/hdlib#7 @anilbey
Hi @cumbof, I noticed one small issue.

The line "This approach recently shown promise in various domains" should be "This approach has recently shown promise in various domains."
