### 2023-09-08 cumbof/hdlib#7 @cumbof
Thanks @anilbey, I'm going to update the manuscript
