### 2023-09-08 4c33fad
Update paper.md
https://github.com/cumbof/hdlib/issues/7#issuecomment-1711534272
