+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-Vector-Symbolic Architectures (VSA, a.k.a. Hyperdimensional Computing) is an emerging computing paradigm that works by combining vectors in a high-dimensional space for representing and processing information [@kanerva2009hyperdimensional; @kanerva2014computing]. This approach recently shown promise in various domains for dealing with different kind of computational problems, including artificial intelligence [@osipov2022hyperseed; @haputhanthri2022evaluating], cognitive science [@gayler2004vector; @graben2022vector], robotics [@neubert2019introduction], natural language processing [@quiroz2020semantic], bioinformatics [@cumbo2020mdpi; @poduval2021cognitive; @chen2022density; @kim2020geniehd], medical informatics [@ni2022neurally; @lagunes2018cancer], cheminformatics [@jones2023hd; @ma2022molehd], and internet of things [@simpkin2020efficient] among other scientific disciplines [@schlegel2022comparison].
+Vector-Symbolic Architectures (VSA, a.k.a. Hyperdimensional Computing) is an emerging computing paradigm that works by combining vectors in a high-dimensional space for representing and processing information [@kanerva2009hyperdimensional; @kanerva2014computing]. This approach has recently shown promise in various domains for dealing with different kind of computational problems, including artificial intelligence [@osipov2022hyperseed; @haputhanthri2022evaluating], cognitive science [@gayler2004vector; @graben2022vector], robotics [@neubert2019introduction], natural language processing [@quiroz2020semantic], bioinformatics [@cumbo2020mdpi; @poduval2021cognitive; @chen2022density; @kim2020geniehd], medical informatics [@ni2022neurally; @lagunes2018cancer], cheminformatics [@jones2023hd; @ma2022molehd], and internet of things [@simpkin2020efficient] among other scientific disciplines [@schlegel2022comparison].
 
 Here we present _hdlib_, a Python library for designing Vector-Symbolic Architectures. Its code is available on GitHub at [https://github.com/cumbof/hdlib](https://github.com/cumbof/hdlib) and it is distributed under the MIT license as a Python package through PyPI (_pip install hdlib_) and Conda on the _conda-forge_ channel (_conda install -c conda-forge hdlib_). GitHub releases are also available on Zenodo at [https://doi.org/10.5281/zenodo.7996502](https://doi.org/10.5281/zenodo.7996502). Documentation with examples of how to use the library is also available at [https://github.com/cumbof/hdlib/wiki](https://github.com/cumbof/hdlib/wiki).
 
