### 2023-09-08 #5704 @cumbof
Another minor fix based on https://github.com/cumbof/hdlib/issues/7#issuecomment-1711534272
