### 2023-09-09 e676743
Update CITATION.cff
