### 2023-09-09 #5704 @cumbof
Thanks @anilbey and @mahfuz05062, I really appreciate your time in reviewing our submission. All your comments have been quite useful for improving code and documentation.

@danielskatz I've just released [hdlib v.0.1.15](https://github.com/cumbof/hdlib/releases/tag/0.1.15) which mainly includes my replies to the reviewers' issues.

Could you please update the package version here as well?

Also, please let me know if there is anything I'm supposed to do now.
Thanks
