### 2023-09-10 #5704 @cumbof
@danielskatz I'm done with the additional author tasks, you can mark them all as completed.
Here is the link to the Zenodo entry https://doi.org/10.5281/zenodo.8331296
