### 2023-09-11 77e8c22
case changes in bib
