+++ paper/paper.bib
@@ -119,7 +119,7 @@
 }
 
 @inproceedings{kim2020geniehd,
-  title={GenieHD: Efficient dna pattern matching accelerator using hyperdimensional computing},
+  title={GenieHD: Efficient {DNA} pattern matching accelerator using hyperdimensional computing},
   author={Kim, Yeseong and Imani, Mohsen and Moshiri, Niema and Rosing, Tajana},
   booktitle={2020 Design, Automation \& Test in Europe Conference \& Exhibition (DATE)},
   pages={115--120},
@@ -181,7 +181,7 @@
 }
 
 @article{gayler2004vector,
-  title={Vector symbolic architectures answer Jackendoff's challenges for cognitive neuroscience},
+  title={Vector symbolic architectures answer {J}ackendoff's challenges for cognitive neuroscience},
   author={Gayler, Ross W},
   journal={arXiv preprint cs/0412059},
   year={2004},
@@ -189,7 +189,7 @@
 }
 
 @article{simpkin2020efficient,
-  title={Efficient orchestration of node-red iot workflows using a vector symbolic architecture},
+  title={Efficient orchestration of {Node-RED IoT} workflows using a vector symbolic architecture},
   author={Simpkin, Chris and Taylor, Ian and Harborne, Daniel and Bent, Graham and Preece, Alun and Ganti, Raghu K},
   journal={Future Generation Computer Systems},
   volume={111},
@@ -200,7 +200,7 @@
 }
 
 @misc{heddes2023torchhd,
-  title={Torchhd: An Open Source Python Library to Support Research on Hyperdimensional Computing and Vector Symbolic Architectures}, 
+  title={Torchhd: An Open Source {P}ython Library to Support Research on Hyperdimensional Computing and Vector Symbolic Architectures}, 
   author={Mike Heddes and Igor Nunes and Pere Vergés and Denis Kleyko and Danny Abraham and Tony Givargis and Alexandru Nicolau and Alexander Veidenbaum},
   year={2023},
   eprint={2205.09208},
@@ -210,7 +210,7 @@
 }
 
 @article{kang2022openhd,
-  title={OpenHD: A gpu-powered framework for hyperdimensional computing},
+  title={OpenHD: A {GPU}-powered framework for hyperdimensional computing},
   author={Kang, Jaeyoung and Khaleghi, Behnam and Rosing, Tajana and Kim, Yeseong},
   journal={IEEE Transactions on Computers},
   volume={71},
