### 2023-09-11 #5704 @danielskatz
👋 @cumbof - I've suggested a bunch of minor fixes for bib entries in https://github.com/cumbof/hdlib/pull/10 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
