### 2023-09-11 #5704 @cumbof
Thanks @danielskatz! I don't have any concerns about your PR. I've just merged it.

Let me know if I can do anything else.

Also, I was wondering if I'm supposed to maintain the `joss` branch in the `hdlib` repo even after the publication.
Since it is out of sync with the main branch, except for the paper folder, I was planning to push the paper in the main branch and get rid of the `joss` branch.

Please let me know if this is possible.
