### 2023-09-11 #5704 @danielskatz
@cumbof - yes, that's fine. We don't have a requirement for a paper branch at all, some submissions have the paper in main from the start.  But please don't do this until after the publication is done.
